Subject: Improved native exception continuations [s48-tuebingen @ 405]
Native exception continuations now have their own layout instead of
being hacked byte-code exception continuations.

This also fixes a bug in the previous version were a code pointer was
pushed on the stack. In addition, when trying to proceed a
non-restartable opcode, the code pointer is initialized to correctly
before raising an exception.