Subject: Don't always enqueue in MAYBE-COMMIT-AND-BLOCK-ON-QUEUE [s48-tuebingen @ 426]
Previously, MAYBE-COMMIT-AND-BLOCK-ON-QUEUE would enqueue something
into its argument whether the commit succeeded (and hence the current
thread actually blocks) or not.  (This would trigger a subtle race
condition bug in PLACEHOLDER-VALUE for which this is, albeit, not the
complete fix.)

So now, MAYBE-COMMIT-AND-BLOCK (which MAYBE-COMMIT-AND-BLOCK-ON-QUEUE
calls) accepts an additional, optional thunk argument which will be
called only in case the commit succeeded.
MAYBE-COMMIT-AND-BLOCK-ON-QUEUE passes a thunk which enqueues the
thread cell.