Subject: First stab at CML library [s48-tuebingen @ 433]
Initial attempt of the CML library: This implements the basic
primitives of Reppy's event algebra, as well as synchronous channels,
placeholders, jars (akin to ID-90) M-structures, and asynchronous
channels.

As Reppy's term "event" seemed a bit generic to me, events are called
"rendezvous" in this library.

Currently, this implements the same API as the CML library in
Sunterlib.  However, the implementation is based on optimistic
concurrency rather than mutual exclusion.