Subject: Get the assembler back to working order [s48-tuebingen @ 505]
This just barely makes the assembler work again.  (It had not been
adapted to the new VM architecture and wouldn't even load.) There's
considerable work to be done.  Specifically, for the LITERAL
instruction to work right, the assembler would need to track the stack
width to get at the template.

However, this is sufficient to get (a new version of)
scheme/misc/annotate.scm to compile.