Subject: Add interrupt checking during calls to native code [s48-tuebingen @ 603]
As discussed on
<http://ml-archive.informatik.uni-tuebingen.de/s48-devel/2003-09/msg00000.html>,
this splits RETURN-FROM-INTERRUPT into 3 instructions,
RETURN-FROM-POLL-INTERRUPT, RESUME-INTERRUPTED-CALL-TO-BYTE-CODE, and
RESUME-INTERRUPTED-CALL-TO-NATIVE-CODE.

POLL and BYTE-CODE work as before. The new code for calls to native
code pushes the closure and the protocol skip on the stack, calls the
interrupt handler in case of an interrupt. The new opcode
RESUME-INTERRUPTED-CALL-TO-NATIVE-CODE pops and calls
s48-call-native-procedure.

This also removes the stack-args argument to CALL-NATIVE-CODE et
al. as it was unused, and the PROC argument to CALL-NATIVE-CODE et
al. as the procedure is now always in *val* to make it visible to the GC.