Subject: Add REINITIALIZERS package [s48-tuebingen @ 717]
A reinitializer is a thunk that runs upon image resumption, and is
conceptually not attached to a record, as record resumers are.  (Even
though this functionality is, of course, implemented by record
resumption.)

Richard originally wrote this for the POSIX signals functionality.  I
just put it into its own package, as the functionality will be needed
in other places as well.