Subject: Don't abort writes without waiters [s48-tuebingen @ 786]
This is a follow-up to rev 749:

ABORT-UNWANTED-I/O would still try to abort writes that had been
initiated by periodic buffer flushing, which work the way one would
expect: The buffer flushing would just re-initiate the aborted write.
This problem isn't visible on Unix, but it is on Windows.

So, for now, change ABORT-UNWANTED-I/O to ABORT-UNWANTED-READS, and
don't abort writes anymore.  In the future, we might indeed abort
writes nobody wants anymore, but that requires larger changes.  (And
the present situation may be undesirable in border cases, but at least
it isn't wrong.)

Also, I've added a comment that explains matters.