Subject: Add IO-ERROR-EVENT [s48-tuebingen @ 878]
Add an IO-ERROR-EVENT to the VM and the interface from the C code and
propagate it to the RTS as a separate I/O-ERROR interrupt.  (Don't do
anything with the interrupt yet, however.)  This doesn't make sense
with the Unix I/O system, but will make sense with the Windows port.