Subject: Add library for creating compact tables [s48-tuebingen @ 934]
The COMPACT-TABLES structure creates one-level compact tables:

A compact table is an encoding of a very large vector that has lots
of recurring patterns.  It was written for encoding Unicode tables.

The vector is partitioned into blocks, and the blocks get assembled
into a new compressed vector.  Each time a new block gets added, the
algorithm looks if the same block is already present in the
compressed vector, or the compressed vector ends with a prefix of
the new block.  In the former case, nothing needs to get added.  In
the latter case, only the suffix needs to get added.  At the same
time, the algorithm computes a table with indices of the block
beginnings.

The algorithm can take a long time; little attempt at optimization
has been made.  It's mainly intended for offline computation as part
of a build process.

I took this from "Unicode Demystified"---I'm guessing the original
idea is probably by Tarjan or someone like him.