Subject: Add GUI to Windows installer [s48-tuebingen @ 989]
This builds upon various bits of sample code off the Internet.  We
even get a graphical confirmation of the license, and "repair" and
"remove" buttons.

There's still a problem in that the installer tries to build
scheme48.image after removal of the package---the InstallMode property
somehow reverts to its old value from one installation phase to the
next.  At least the resulting error message is harmless.