Subject: Fix bug in / opcode [s48-tuebingen @ 1278]
Run:

,bench
,open cells
(define (f)
  (/ (begin
       (make-cell 1)
       2130612)
     1152921504606846976)
  (f))
(f)

Boom.

The reason is that, if / operates on bignums and produces a non-zero
remainder, it needs to pass its arguments to the exception handler.
However, finding out may involve a GC, which means we need to make
sure to pass the arguments *after* GC, not the stale ones from before.