Subject: Convert FILE-OPTIONS to enum sets [s48-tuebingen @ 1358]
(The C side of this patch was accidentally committed as part of rev
1357.)

The FILE-OPTIONS type seems to've been implemented before enum sets
were added; implement in terms of enum sets.

Also, add FILE-OPTIONS-UNION, so I can programmatically construct
FILE-OPTIONS objects.  (As I've had to do for the reference
implementation of SRFI 68.)