Subject: Add implementation of SRFI 62: S-expression comments [s48-tuebingen @ 1384]
This modifies the reader permanently when loaded, which is bad.  On
the other hand, it will probably be in R6RS, so the problem should go
away by itself.

Note this is slightly half-baked: it will work in programs, but at the
prompt,

> #;(ignore me)

doesn't give you another prompt---it blocks waiting for the next form.