Subject: Make VERIFY-LATER! really verify later [s48-tuebingen @ 1561]
The usage of VERIFY-LATER! in REALLY-DEF (in module-usage.scm) assumes
that the verification will run once the bindings established by DEFINE
are complete.  The actual definition provided in pedit.scm ran the
thunk immediately, though.  This would mean that the previous binding
or (in the case of DEFINE-STRUCTURES) #f would get verified.

To fix, delay the running of the thunk to the actual verification.

Reported by: Taylor Campbell