Subject: Replace JUMP-IF-NOT-... by TEST-OP+JUMP-IF-FALSE [s48-tuebingen @ 2122]
The previous JUMP-IF-NOT-... instructions required severe
implementation effort in the VM. Replace them by
TEST-OP+JUMP-IF-FALSE, which takes an offset plus two fully functional
instructions as arguments. The offset is the jump target relative to
the PC of TEST-OP+JUMP-IF-FALSE, the first instruction does the test,
the second is JUMP-IF-FALSE.

The VM implementation can now simply skip over to the first
instruction. The native-code compiler can extract the jump target from
the first argument and fuse the two instructions into a test and a
conditional jump.