Fix potential crash after current-cont opcode.
The "current-cont" opcode uses "copy-current-continuation-to-heap" to
store the current continuation on the heap. It adjusts *cont* but not
*stack*. Triggering a garbage collection in a subsequent opcode like
"make-flat-env" can cause a crash:

The GC traces the upper part of the stack

(s48-trace-locations! *stack*
                      (address+ *stack*
                                (cells->a-units (operands-on-stack))))

(define (operands-on-stack)
  (a-units->cells (address-difference *cont* *stack*)))

that now includes the *whole* stack. However, there can be structures
on the stack the GC cannot handle correctly.

With the patch, *stack* should be adjusted properly.

Thanks to Marcel Turino for finding and fixing this.